Simplify the code base by dropping support for the manifest file
Initially, I had a plan to keep a backup of specific versions and store
them separately on the CDN. For this purpose, I came up with the idea of
using a version manifest file, with links to download a specific version.
But later, rethinking the strategy of this project, I decided that this
would be excessive and would probably lead to an excessive amount of
support for the infrastructure.